% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited-methods.R
\name{Dog}
\alias{Dog}
\title{super class for all dogs}
\description{
super class for all dogs

super class for all dogs
}
\section{Super class}{
\code{\link[R6test:Animal]{R6test::Animal}} -> \code{Dog}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bark}{\code{Dog$bark()}}
\item \href{#method-clone}{\code{Dog$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6test" data-topic="Animal" data-id="initialize">}\href{../../R6test/html/Animal.html#method-initialize}{\code{R6test::Animal$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bark"></a>}}
\if{latex}{\out{\hypertarget{method-bark}{}}}
\subsection{Method \code{bark()}}{
the noise made by dogs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dog$bark()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
